'CausalQueries'

'CausalQueries' is a package that lets users generate binary causal models, update over models given data, and calculate arbitrary causal queries. Model definition makes use of dagitty type syntax. Updating is implemented in 'stan'.

See also

Author

Maintainer: Till Tietz ttietz2014@gmail.com (ORCID)

Authors:

Macartan Humphreys macartan@gmail.com (ORCID)

Alan Jacobs alan.jacobs@ubc.ca

Lily Medina lilymiru@gmail.com (ORCID)

Georgiy Syunyaev georgiy.syunyaev@vanderbilt.edu (ORCID)

Other contributors:

Clara Bicalho clarabmcorreia@gmail.com [contributor]

Jasper Cooper jjc2247@columbia.edu [contributor]

Merlin Heidemanns mnh2123@columbia.edu [contributor]

Julio Solis juliosolisar@gmail.com [contributor]