Create parameter documentation to inherit

Create parameter documentation to inherit

Usage

CausalQueries_internal_inherit_params(
  model,
  query,
  join_by,
  parameters,
  P,
  A,
  data,
  data_events,
  node,
  statement,
  using,
  n_draws
)

Arguments

model

A causal_model. A model object generated by make_model.

query

A character string. An expression defining nodal types to interrogate realise_outcomes. An expression of the form "Y[X=1]" asks for the value of Y when X is set to 1

join_by

A logical operator. Used to connect causal statements: AND ('&') or OR ('|'). Defaults to '|'.

parameters

A vector of real numbers in [0,1]. Values of parameters to specify (optional). By default, parameters is drawn from the parameters dataframe. See inspect(model, "parameters_df").

P

A data.frame. Parameter matrix. Not required but may be provided to avoid repeated computation for simulations. See inspect(model, "parameter_matrix").

A

A data.frame. Ambiguities matrix. Not required but may be provided to avoid repeated computation for simulations. inspect(model, "ambiguities_matrix")

data

A data.frame. Data of nodes that can take three values: 0, 1, and NA. In long form as generated by make_events

data_events

A data.frame. It must be compatible with nodes in model. The default columns are event, strategy and count.

node

A character string. The quoted name of a node.

statement

A character string. A quoted causal statement.

using

A character string. Indicates whether to use `priors`, `posteriors` or `parameters`.

n_draws

An integer. If no prior distribution is provided, generate prior distribution with n_draws number of draws.

Value

This function does not return anything. It is used to inherit roxygen documentation