Helper to fill in missing do operators in causal expression

Helper to fill in missing do operators in causal expression

Usage

add_dots(q, model)

Arguments

q

A character string. Causal query with at least one parent node missing their do operator.

model

A causal_model. A model object generated by make_model.

Value

A causal query expression with all parents nodes set to either 0, 1 or wildcard '.'.

Examples

# \donttest{
model <- make_model('X -> Y <- M')
CausalQueries:::add_dots('Y[X=1]', model)
#> [1] "Y[X=1, M = . ]"
CausalQueries:::add_dots('Y[]', model)
#> [1] "Y[M = . , X = . ]"
# }