helper to get types from queries

helper to get types from queries

Usage

queries_to_types(jobs, model, query_col, realisations)

Arguments

jobs

a data frame of argument combinations

model

a list of models

query_col

string specifying the name of the column in jobs holding queries to be evaluated

realisations

list of data frame outputs from calls to realise_outcomes

Value

jobs data frame with a nested column of map_query_to_nodal_type outputs