Reduce nodal types using labels

Reduce nodal types using labels

Usage

restrict_by_labels(model, labels, given = NULL, keep = FALSE)

Arguments

model

A causal_model. A model object generated by make_model.

labels

A list of character vectors specifying nodal types to be kept or removed from the model.

given

A character vector or list of character vectors specifying nodes on which the parameter set to be restricted depends. When mixing labels that are further restricted by given and ones that are not, labels without given restrictions should have given specified as one of NULL, NA, "" or " ".

keep

Logical. If `FALSE`, removes and if `TRUE` keeps only causal types specified by restriction.

Value

A list with two components: 1. a vector with parameters names of parameters implicated by the restrictions, 2. a vector of subsetting instructions used to identify implicated causal types

See also