Generate estimands dataframe

Calculated from a parameter vector, from a prior or from a posterior distribution.

Usage

query_model(
  model,
  queries = NULL,
  given = NULL,
  using = list("parameters"),
  parameters = NULL,
  stats = NULL,
  n_draws = 4000,
  expand_grid = FALSE,
  case_level = FALSE,
  query = NULL,
  cred = 95
)

Arguments

model

A causal_model. A model object generated by make_model.

queries

A vector of strings or list of strings specifying queries on potential outcomes such as "Y[X=1] - Y[X=0]".

given

A vector or list of strings specifying givens. A given is a quoted expression that evaluates to a logical statement. Allows estimand to be conditioned on *observational* (or counterfactual) distribution.

using

A vector or list of strings. Whether to use priors, posteriors or parameters.

parameters

A vector of real numbers in [0,1]. Values of parameters to specify (optional). By default, parameters is drawn from model$parameters_df.

stats

Functions to be applied to estimand distribution. If NULL, defaults to mean, standard deviation, and 95% confidence interval. Functions should return a single numeric value.

n_draws

An integer. Number of draws.

expand_grid

Logical. If TRUE then all combinations of provided lists are examined. If not then each list is cycled through separately. Defaults to FALSE.

case_level

Logical. If TRUE estimates the probability of the query for a case.

query

alias for queries

cred

size of the credible interval ranging between 0 and 100

Value

A DataFrame with columns Model, Query, Given and Using defined by corresponding input values. Further columns are generated as specified in stats.

Details

Queries can condition on observed or counterfactual quantities. Nested or "complex" counterfactual queries of the form Y[X=1, M[X=0]] are allowed.

Examples

model <- make_model("X -> Y") %>% set_prior_distribution(n_draws = 10000)

# \donttest{

# `expand_grid= TRUE` requests the Cartesian product of arguments

models <- list(
 M1 = make_model("X -> Y"),
 M2 = make_model("X -> Y") |> set_restrictions("Y[X=1] < Y[X=0]")
 )

query_model(
  models,
  query = list(ATE = "Y[X=1] - Y[X=0]",
               Share_positive = "Y[X=1] > Y[X=0]"),
  given = c(TRUE,  "Y==1 & X==1"),
  using = c("parameters", "priors"),
  expand_grid = FALSE)
#>   model          query       given      using case_level      mean        sd
#> 1    M1            ATE           - parameters      FALSE 0.0000000        NA
#> 2    M1 Share_positive Y==1 & X==1     priors      FALSE 0.4967603 0.2861068
#> 3    M2            ATE           - parameters      FALSE 0.3333333        NA
#> 4    M2 Share_positive Y==1 & X==1     priors      FALSE 0.4988770 0.2893070
#>     cred.low cred.high
#> 1 0.00000000 0.0000000
#> 2 0.02473718 0.9735785
#> 3 0.33333333 0.3333333
#> 4 0.02603898 0.9733474

query_model(
  models,
  query = list(ATE = "Y[X=1] - Y[X=0]",
               Share_positive = "Y[X=1] > Y[X=0]"),
  given = c(TRUE,  "Y==1 & X==1"),
  using = c("parameters", "priors"),
  expand_grid = TRUE)
#>    model          query       given      using case_level        mean        sd
#> 1     M1            ATE           - parameters      FALSE 0.000000000        NA
#> 2     M2            ATE           - parameters      FALSE 0.333333333        NA
#> 3     M1            ATE           -     priors      FALSE 0.001656862 0.3211220
#> 4     M2            ATE           -     priors      FALSE 0.334258094 0.2395079
#> 5     M1            ATE Y==1 & X==1 parameters      FALSE 0.500000000        NA
#> 6     M2            ATE Y==1 & X==1 parameters      FALSE 0.500000000        NA
#> 7     M1            ATE Y==1 & X==1     priors      FALSE 0.504624275 0.2897680
#> 8     M2            ATE Y==1 & X==1     priors      FALSE 0.497620808 0.2909957
#> 9     M1 Share_positive           - parameters      FALSE 0.250000000        NA
#> 10    M2 Share_positive           - parameters      FALSE 0.333333333        NA
#> 11    M1 Share_positive           -     priors      FALSE 0.251916852 0.1962484
#> 12    M2 Share_positive           -     priors      FALSE 0.334258094 0.2395079
#> 13    M1 Share_positive Y==1 & X==1 parameters      FALSE 0.500000000        NA
#> 14    M2 Share_positive Y==1 & X==1 parameters      FALSE 0.500000000        NA
#> 15    M1 Share_positive Y==1 & X==1     priors      FALSE 0.504624275 0.2897680
#> 16    M2 Share_positive Y==1 & X==1     priors      FALSE 0.497620808 0.2909957
#>        cred.low cred.high
#> 1   0.000000000 0.0000000
#> 2   0.333333333 0.3333333
#> 3  -0.650436869 0.6402530
#> 4   0.012679481 0.8439040
#> 5   0.500000000 0.5000000
#> 6   0.500000000 0.5000000
#> 7   0.027161678 0.9750710
#> 8   0.025131825 0.9769851
#> 9   0.250000000 0.2500000
#> 10  0.333333333 0.3333333
#> 11  0.009580428 0.7075848
#> 12  0.012679481 0.8439040
#> 13  0.500000000 0.5000000
#> 14  0.500000000 0.5000000
#> 15  0.027161678 0.9750710
#> 16  0.025131825 0.9769851

# An example of a custom statistic: uncertainty of token causation
f <- function(x) mean(x)*(1-mean(x))

query_model(
  model,
  using = list( "parameters", "priors"),
  query = "Y[X=1] > Y[X=0]",
  stats = c(mean = mean, sd = sd, token_variance = f))
#>             query given      using case_level      mean        sd
#> 1 Y[X=1] > Y[X=0]     - parameters      FALSE 0.2500000        NA
#> 2 Y[X=1] > Y[X=0]     -     priors      FALSE 0.2479113 0.1944187
#>   token_variance
#> 1      0.1875000
#> 2      0.1864513
# }