Summarizing causal models

summary method for class "causal_model".

Usage

# S3 method for class 'causal_model'
summary(object, ...)

# S3 method for class 'summary.causal_model'
print(x, include = NULL, ...)

Arguments

object

An object of causal_model class produced using make_model or update_model.

...

Further arguments passed to or from other methods.

x

An object of summary.causal_model class, produced using summary.causal_model.

include

A character string specifying the objects summaries to print. Defaults to NULL printing causal statement, specification of nodal types and summary of model restrictions. See Details for full list of available values.

Value

Returns the object of class summary.causal_model that preserves the list structure of causal_model class and adds the following additional elements:

"parents" a list of parents of all nodes in a model,

"parameters" a vector of 'true' parameters,

"parameter_names" a vector of names of parameters,

"parameter_mapping" a matrix mapping from parameters into data types,

"parameter_matrix" a matrix mapping from parameters into causal types,

"causal_types" a data frame listing causal types and the nodal types that produce them,

"data_types" a list with the all data types consistent with the model; for options see "?get_all_data_types",

"prior_event_probabilities" a vector of prior data (event) probabilities given a parameter vector; for options see "?get_event_probabilities",

"ambiguities_matrix" a matrix mapping from causal types into data types,

"prior_hyperparameters" a vector of alpha values used to parameterize Dirichlet prior distributions; optionally provide node names to reduce output "inspect(prior_hyperparameters, c('M', 'Y'))"

Details

print.summary.causal_model reports causal statement, full specification of nodal types and summary of model restrictions. By specifying `include` argument users can instead print a custom summary of any set of the following objects contained in the `summary.causal_model`:

"statement" A character string giving the causal statement,

"nodes" A list containing the nodes in the model,

"parents" A list of parents of all nodes in a model,

"parents_df" A data frame listing nodes, whether they are root nodes or not, and the number and names of parents they have,

"parameters" A vector of 'true' parameters,

"parameters_df" A data frame containing parameter information,

"parameter_names" A vector of names of parameters,

"parameter_mapping" A matrix mapping from parameters into data types,

"parameter_matrix" A matrix mapping from parameters into causal types,

"causal_types" A data frame listing causal types and the nodal types that produce them,

"nodal_types" A list with the nodal types of the model,

"data_types" A list with the all data types consistent with the model; for options see `"?get_all_data_types"`,

"prior_hyperparameters" A vector of alpha values used to parameterize Dirichlet prior distributions; optionally provide node names to reduce output `inspect(prior_hyperparameters, c('M', 'Y'))`

"prior_distribution" A data frame of the parameter prior distribution,

"prior_event_probabilities" A vector of data (event) probabilities given a single realization of parameters; for options see `"?get_event_probabilities"`,

"ambiguities_matrix" A matrix mapping from causal types into data types,

"type_prior" A matrix of type probabilities using priors,

"type_distribution" A matrix of type probabilities using posteriors,

"posterior_distribution" A data frame of the parameter posterior distribution,

"posterior_event_probabilities" A sample of data (event) probabilities from the posterior,

"data" A data frame with data that was used to update model.

"stanfit" A `stanfit` object generated by Stan,

"stan_summary" A `stanfit` summary with updated parameter names,

"stan_objects" A list of Stan outputs that includes `stanfit`, `data`, and, if requested when updating the model, posterior `event_probabilities` and `type_distribution`,

Examples

# \donttest{

model <-
  make_model('X -> Y') |>
  update_model(
    keep_event_probabilities = TRUE,
    keep_fit = TRUE,
    data = make_data(model, n = 100))
#> Error: object 'model' not found

summary(model)
#> Error: object 'model' not found

# }

# \donttest{
model <-
  make_model('X -> Y')

print(summary(model), include = "type_distribution")
#> Error in print.summary.causal_model(summary(model), include = "type_distribution"): Model does not contain the following requested objects: posterior type_distribution;
#>   to include these objects update model with 'type_distribution = TRUE'
print(summary(model), include = "posterior_distribution")
#> Error in print.summary.causal_model(summary(model), include = "posterior_distribution"): Model does not contain: posterior_distribution;
#>   to include these objects update model
print(summary(model), include = "posterior_event_probabilities")
#> Error in print.summary.causal_model(summary(model), include = "posterior_event_probabilities"): Model does not contain the following requested objects: event_probabilities;
#>   to include these objects update model with 'keep_event_probabilities = TRUE'
print(summary(model), include = "data_types")
#> 
#> Data frame of all possible data (events) given the model:
#> 
#>      event  X  Y
#> X0Y0  X0Y0  0  0
#> X1Y0  X1Y0  1  0
#> X0Y1  X0Y1  0  1
#> X1Y1  X1Y1  1  1
#> Y0      Y0 NA  0
#> Y1      Y1 NA  1
#> X0      X0  0 NA
#> X1      X1  1 NA
#> None  None NA NA
#> 
print(summary(model), include = "ambiguities_matrix")
#> 
#> Mapping from causal types into data types:
#> 
#>       X0Y0 X1Y0 X0Y1 X1Y1
#> X0Y00    1    0    0    0
#> X1Y00    0    1    0    0
#> X0Y10    0    0    1    0
#> X1Y10    0    1    0    0
#> X0Y01    1    0    0    0
#> X1Y01    0    0    0    1
#> X0Y11    0    0    1    0
#> X1Y11    0    0    0    1
#> 
print(summary(model), include = "prior_hyperparameters")
#> 
#> Alpha parameter values used for Dirichlet prior distributions:
#> 
#>  X.0  X.1 Y.00 Y.10 Y.01 Y.11 
#>    1    1    1    1    1    1 
#> 

model <-
  model |>
  update_model(
    keep_event_probabilities = TRUE,
    keep_fit = TRUE,
    data = simulate_data(model, n = 100))
#> Error in simulate_data(model, n = 100): could not find function "simulate_data"

print(summary(model), include = c("statement", "nodes"))
#> 
#> Causal statement: 
#> X -> Y
#> 
#> Nodes: 
#> X, Y
#> 
print(summary(model), include = "parameters_df")
#> 
#> Mapping of model parameters to nodal types: 
#> 
#>   param_names: name of parameter
#>   node:        name of endogeneous node associated
#>                with the parameter
#>   gen:         partial causal ordering of the
#>                parameter's node
#>   param_set:   parameter groupings forming a simplex
#>   given:       if model has confounding gives
#>                conditioning nodal type
#>   param_value: parameter values
#>   priors:      hyperparameters of the prior
#>                Dirichlet distribution 
#> 
#>   param_names node gen param_set nodal_type given param_value priors
#> 1         X.0    X   1         X          0              0.50      1
#> 2         X.1    X   1         X          1              0.50      1
#> 3        Y.00    Y   2         Y         00              0.25      1
#> 4        Y.10    Y   2         Y         10              0.25      1
#> 5        Y.01    Y   2         Y         01              0.25      1
#> 6        Y.11    Y   2         Y         11              0.25      1
#> 
print(summary(model), include = "posterior_event_probabilities")
#> Error in print.summary.causal_model(summary(model), include = "posterior_event_probabilities"): Model does not contain the following requested objects: event_probabilities;
#>   to include these objects update model with 'keep_event_probabilities = TRUE'
print(summary(model), include = "posterior_distribution")
#> Error in print.summary.causal_model(summary(model), include = "posterior_distribution"): Model does not contain: posterior_distribution;
#>   to include these objects update model
print(summary(model), include = "data")
#> Error in print.summary.causal_model(summary(model), include = "data"): Model does not contain the following requested objects: specified 'data';
#>   to include these objects update model with 'data'
print(summary(model), include = "stanfit")
#> Error in print.summary.causal_model(summary(model), include = "stanfit"): Model does not contain the following requested objects: stanfit;
#>   to include these objects update model with 'keep_fit = TRUE'
print(summary(model), include = "type_distribution")
#> Error in print.summary.causal_model(summary(model), include = "type_distribution"): Model does not contain the following requested objects: posterior type_distribution;
#>   to include these objects update model with 'type_distribution = TRUE'
print(summary(model), include = "stan_objects")
#> Error in print.summary.causal_model(summary(model), include = "stan_objects"): Model does not contain: stan_objects;
#>   to include these objects update model
# }